Version 1.3.2 ~ Build 1.8.3 Large Bugfix Update :3
This update includes the following changes and additions:
~ Unpatching of the Instance of the `LazyActionSheet` (Memory Leak Fix)
This means that all state hooks had to be converted to let variables.
This also means that the `ActionSheet` will rerender less than before.
~ Remove all languages which aren't valid in the Google Translate Engine
(but which are entries in the `ISO-639-1/2T`)
~ Rewrite entire `Translating Funtionality` by hand and convert to TS
~ Rename the `Utils` folder to `Common` instead.
~ Add a new option to the </dislate:1> command: `Example`
This command allows you to send an example message in a random language.
~ Rewrite execution of commands to promises
This means you can reply to people with /translate and /dislate debug
~ Convert the ISO-639-1 Names list and ISO-639-2T List to single arrays
This means that the key value pair can be constructed in code.
Therefore, the build output is now smaller by a few KBs.
~ The Settings page and any children components now use
the constructed `names` and `iso` object from index.tsx as a prop.
~ As the ISO-2/ISO-1 object can now be reversed to begin with,
The `Format.reverse` function is no longer needed and has been removed.

Overall, all users are recommended to update to this version.