Version 1.3.6 ~ Build 2.1.6
Changes and additions:
~ Change Debug Info page to render as an ActionSheet
This also comes with the addition of being able to:
    ~ Open the page from Plugin Settings
    ~ Clear/Disable all options
~ The buttons to Send/Copy the log have been moved under the search bar.
~ Change the Shadow implementation to be a method rather than an object
This was done to allow to set an optional opacity on the shadow.
~ Change the stateStore to use Tthe Dislate Store rather than Storage.
This was done because Storage was not setting the items correctly
(This also means that the Clear Stores button now works correctly :3)
~ Modifed the Debug Command to no longer push a page to Navigation
Instead, open the LazyActionSheet.
~ Also added refreshing to the Languages page.
~ Dialog now takes up 2% less space to fit with the other components.
~ Removed the Page component because nothing is using it anymore
(Also removed the ExitWrapper component at the same time)
(This was done in favour of ActionSheet, neither are required anymore)
~ Modify the Debug Info option in settings to open the ActionSheet
(instead of copying the entire log to clipboard :3)